+++ example/a.ts
@@ -1,7 +1,3 @@
 import { Elysia, t } from '../src'
 
-using plugin = new Elysia().get('a', 'a')
-
-const main = new Elysia().use(plugin)
-
-main.listen(3000)
+const main = new Elysia().get('/', 'hello').listen(3000)