+++ example/stress/multiple-routes.ts
@@ -1,9 +1,10 @@
 import { Elysia, t } from '../../src'
+import v8 from 'v8'
 
 const total = 500
 
 {
-	const app = new Elysia()
+	const app = new Elysia({ precompile: true })
 
 	const t1 = performance.now()
 	const memory = process.memoryUsage().heapTotal / 1024 / 1024
@@ -26,4 +27,6 @@ const total = 500
 
 	console.log(memoryAfter - memory, 'MB memory used')
 	console.log(((memoryAfter - memory) / total) * 1024, 'KB memory used')
+
+	// v8.writeHeapSnapshot()
 }